requirements.txt
streamlit
yt-dlp
requests
webvtt-py
sumy
nltk